+++ src/index.tsx
@@ -1,0 +1,69 @@
+export {
+  usePosts,
+  useCreatePost,
+  useRetrivePost,
+  useUpdatePost,
+  useDeletePost
+} from './hooks/usePosts';
+
+export {
+  useRevisions,
+  useRetriveRevision,
+  useDeleteRevision
+} from './hooks/useRevisions';
+
+export {
+  useCategories,
+  useCreateCategory,
+  useRetriveCategory,
+  useUpdateCategory,
+  useDeleteCategory
+} from './hooks/useCategories';
+
+export {
+  useTags,
+  useCreateTag,
+  useRetriveTag,
+  useUpdateTag,
+  useDeleteTag
+} from './hooks/useTags';
+
+export {
+  usePages,
+  useCreatePage,
+  useRetrivePage,
+  useUpdatePage,
+  useDeletePage
+} from './hooks/usePages';
+
+export {
+  useComments,
+  useCreateComment,
+  useRetriveComment,
+  useUpdateComment,
+  useDeleteComment
+} from './hooks/useComments';
+
+export { useTaxonomies, useRetriveTaxonomy } from './hooks/useTaxonomies';
+
+export {
+  useMedia,
+  useCreateMedia,
+  useRetriveMedia,
+  useUpdateMedia,
+  useDeleteMedia
+} from './hooks/useMedia';
+
+export {
+  useUsers,
+  useCreateUser,
+  useRetriveUser,
+  useUpdateUser,
+  useDeleteUser
+} from './hooks/useUsers';
+
+export { useTypes, useRetriveType } from './hooks/useTypes';
+export { useStatuses, useRetriveStatus } from './hooks/useStatuses';
+export { useSettings, useUpdateSetting } from './hooks/useSettings';
+
+export { WPContext } from './context';