+++ CHANGELOG.md
@@ -1,1 +1,20 @@
 # react-wordpress-hooks Changelog
+
+## 1.0.0 alpha 1 (Soon)
+#### New Feature
+**context**
+- `WPContext` with url value
+
+**hooks, hooks, hooks ...moreHooks**
+- hooks for Posts: `usePosts`, `useCreatePost`, `useRetrivePost`, `useUpdatePost`, `useDeletePost`
+- hooks for Revisions: `useRevisions`, `useRetriveRevision`, `useDeleteRevision`
+- hooks for Categories: `useCategories`, `useCreateCategory`, `useRetriveCategory`, `useUpdateCategory`, `useDeleteCategory`
+- hooks for Tags: `useTags`, `useCreateTag`, `useRetriveTag`, `useRetriveTag`, `useDeleteTag`
+- hooks for Pages: `usePages`, `useCreatePage`, `useRetrivePage`, `useUpdatePage`, `useDeletePage`
+- hooks for Comments: `useComments`, `useCreateComment`, `useRetriveComment`, `useUpdateComment`, `useDeleteComment`
+- hooks for Taxonomies: `useTaxonomies`, `useRetriveTaxonomy`
+- hooks for Media: `useMedia`, `useCreateMedia`, `useRetriveMedia`, `useUpdateMedia`, `useDeleteMedia`
+- hooks for Users: `useUsers`, `useCreateUser`, `useRetriveUser`, `useUpdateUser`, `useDeleteUser`
+- hooks for Types: `useTypes`, `useRetriveType`
+- hooks for Statuses: `useStatuses`, `useRetriveStatus`
+- hooks for Settings: `useSettings`, `useUpdateSetting`